dpll: fix dpll_pin_alloc
The dpll_init_pin function was added for the same reason as
dpll_pin_alloc, allow external module to allocate and init a dpll_pin.
Merge two functions.
dpll_pin_alloc function was used and fixed, its name and behavior is
consistent with the function which allocates dpll, dpll_init_pin was
removed as it was redundant.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>